Add handling of quotes to deleted posts that we managed to restore from archives. Now such posts with be marked with (DEAD) suffix. Also fix issues with the last used archive. The problem was that we only stored one archives which was incorrect because if disabling that archive we would still use that archive as the last used in ArchivesManager.getArchiveDescriptor() method. Now we store all archives that were used with particular thread and only the latest not disabled archive is returned.

+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/core/presenter/ThreadPresenter.kt
@@ -1188,7 +1188,7 @@ class ThreadPresenter @Inject constructor(
         )
 
         val threadLoadable = databaseManager.databaseLoadableManager.get(loadable)
-        threadLoadable.markedNo = threadLink.postId
+        threadLoadable.markedNo = threadLink.postId.toInt()
 
         threadPresenterCallback?.showThread(threadLoadable)
       }
